Really hacky prototype of Task.whenSelected
When it comes to lazy configuration, a really valuable heuristic
is to "only do work if a certain task is selected to run".

Normal lazy configuration relies on "was my task inadvertantly created";
there are many unfortunate but valid reasons to eagerly realize a task,
yet there is still potentially expensive work that you might need to run
_just before_ the task graph is "filled but not yet frozen".

Task.whenSelected is the "last round of user code able to create tasks".
It currently drains all whenSelected callbacks of all tasks in a loop,
so it would be perfectly normal and valid to submit tasks via any Task.

The impementation here may be a little naive,
but the idea behind it, I think serves an important need.

Tests forthcoming.