Add MultiCapabilitySoftwareComponent
Enables a SoftwareComponent to explicitly publish
multiple capabilities.  This allows the publishing
tools to declare what capabilities it will publish,
so those capabilities can be mapped to by coordinate.

This was done before feature variants was released,
so may not be strictly necessary anymore (though it
appears to be so, for the sake of publishing)